requirements.txt
Flask
requests
PyJWT>=2.0.0
cryptography>=3.4.0
rich>=10.0.0
python-dotenv
pandas
tabulate
PyYAML 